requirements.txt
numpy==1.17.3
Pillow==6.2.0
six==1.12.0
torch==1.3.0
torchvision==0.4.1
tqdm==4.36.1
